backend/requirements.txt
flask==2.0.1
flask-mysqldb==0.2.0
gunicorn==20.1.0
mysqlclient==2.1.0
PyMySQL==1.0.2